Split base class from detector-specific subclasses
